Mejoras generales en el servidor intermedio: * Ahora se puede limitar el tamaño de los ficheros. * Se borran los ficheros caducados tanto en almacén como en recuperación. * Solo se borran los ficheros propios, no se tocan los ajenos.

+++ fire-signature/src/main/java/es/gob/fire/server/services/storage/RetrieveService.java
@@ -24,9 +24,10 @@ import javax.servlet.http.HttpServletResponse;
 
 import es.gob.fire.signature.ConfigManager;
 
-/** Servicio de almacenamiento temporal de firmas. &Uacute;til para servir de intermediario en comunicaci&oacute;n
- * entre JavaScript y <i>Apps</i> m&oacute;viles nativas.
- * @author Tom&aacute;s Garc&iacute;a-;er&aacute;s */
+/** Servicio de almacenamiento temporal de datos.
+ * &Uacute;til para servir de intermediario en comunicaci&oacute;n
+ * entre JavaScript y aplicaciones nativas.
+ * @author Tom&aacute;s Garc&iacute;a-Mer&aacute;s. */
 public final class RetrieveService extends HttpServlet {
 
 	private static final long serialVersionUID = -3272368448371213403L;
@@ -48,48 +49,48 @@ public final class RetrieveService extends HttpServlet {
 	@Override
 	protected void service(final HttpServletRequest request, final HttpServletResponse response) throws ServletException, IOException {
 
-		LOGGER.fine("== INICIO DE LA RECUPERACION =="); //$NON-NLS-1$
-
 		final String operation = request.getParameter(PARAMETER_NAME_OPERATION);
 		final String syntaxVersion = request.getParameter(PARAMETER_NAME_SYNTAX_VERSION);
 		response.setHeader("Access-Control-Allow-Origin", "*"); //$NON-NLS-1$ //$NON-NLS-2$
 		response.setContentType("text/plain"); //$NON-NLS-1$
 		response.setCharacterEncoding("utf-8"); //$NON-NLS-1$
 
-		final PrintWriter out = response.getWriter();
-		if (operation == null) {
-			LOGGER.warning(ErrorManager.genError(ErrorManager.ERROR_MISSING_OPERATION_NAME));
-			out.println(ErrorManager.genError(ErrorManager.ERROR_MISSING_OPERATION_NAME));
-			out.flush();
-			return;
-		}
-		if (syntaxVersion == null) {
-			LOGGER.warning(ErrorManager.genError(ErrorManager.ERROR_MISSING_SYNTAX_VERSION));
-			out.println(ErrorManager.genError(ErrorManager.ERROR_MISSING_SYNTAX_VERSION));
-			out.flush();
-			return;
-		}
+		try (
+			final PrintWriter out = response.getWriter();
+		) {
+			if (operation == null) {
+				LOGGER.warning(ErrorManager.genError(ErrorManager.ERROR_MISSING_OPERATION_NAME));
+				out.println(ErrorManager.genError(ErrorManager.ERROR_MISSING_OPERATION_NAME));
+				out.flush();
+				return;
+			}
+			if (syntaxVersion == null) {
+				LOGGER.warning(ErrorManager.genError(ErrorManager.ERROR_MISSING_SYNTAX_VERSION));
+				out.println(ErrorManager.genError(ErrorManager.ERROR_MISSING_SYNTAX_VERSION));
+				out.flush();
+				return;
+			}
 
-		if (OPERATION_RETRIEVE.equalsIgnoreCase(operation)) {
-			retrieveSign(out, request);
+			if (OPERATION_RETRIEVE.equalsIgnoreCase(operation)) {
+				retrieveSign(out, request);
+			}
+			else {
+				LOGGER.warning(ErrorManager.genError(ErrorManager.ERROR_UNSUPPORTED_OPERATION_NAME));
+				out.println(ErrorManager.genError(ErrorManager.ERROR_UNSUPPORTED_OPERATION_NAME));
+			}
+			out.flush();
 		}
-		else {
-			LOGGER.warning(ErrorManager.genError(ErrorManager.ERROR_UNSUPPORTED_OPERATION_NAME));
-			out.println(ErrorManager.genError(ErrorManager.ERROR_UNSUPPORTED_OPERATION_NAME));
-		}
-		out.flush();
-		LOGGER.fine("== FIN DE LA RECUPERACION =="); //$NON-NLS-1$
 
 		// Antes de salir revisamos todos los ficheros y eliminamos los caducados.
-		LOGGER.fine("Limpiamos el directorio temporal"); //$NON-NLS-1$
-		removeExpiredFiles();
+		StorageConfig.removeExpiredFiles();
 	}
 
-	/** Recupera la firma del servidor.
-	 * @param response Respuesta a la petici&oacute;n.
+	/** Recupera los datos del servidor.
+	 * @param out Respuesta a la petici&oacute;n.
 	 * @param request Petici&oacute;n.
 	 * @throws IOException Cuando ocurre un error al general la respuesta. */
-	private static void retrieveSign(final PrintWriter out, final HttpServletRequest request) throws IOException {
+	private static void retrieveSign(final PrintWriter out,
+			                         final HttpServletRequest request) throws IOException {
 
 		final String id = request.getParameter(PARAMETER_NAME_ID);
 		if (id == null) {
@@ -100,11 +101,11 @@ public final class RetrieveService extends HttpServlet {
 
 		LOGGER.fine("Se solicita el fichero con el identificador: " + id); //$NON-NLS-1$
 
-		final File inFile = new File(ConfigManager.getAfirmaTempDir(), id);
+		final File inFile = new File(ConfigManager.getAfirmaTempDir(), StorageConfig.FILE_PREFIX + id);
 
 		// No hacemos distincion si el archivo no existe, no es un fichero, no puede leerse o ha caducado
 		// para evitar que un atacante conozca su situacion. Lo borramos despues de usarlo
-		if (!inFile.exists() || !inFile.isFile() || !inFile.canRead() || isExpired(inFile, ConfigManager.getAfirmaTempsTimeout())) {
+		if (!inFile.isFile() || !inFile.canRead() || StorageConfig.isExpired(inFile, ConfigManager.getAfirmaTempsTimeout())) {
 
 			if (!inFile.exists()) {
 				LOGGER.warning("El fichero con el identificador '" + id + "' no existe: " + inFile.getAbsolutePath()); //$NON-NLS-1$ //$NON-NLS-2$
@@ -124,56 +125,28 @@ public final class RetrieveService extends HttpServlet {
 			);
 			// Que el fichero sea de tipo fichero, implica que existe
 			if (inFile.isFile()) {
-				inFile.delete();
+				if (!StorageConfig.DEBUG) {
+					inFile.delete();
+				}
 			}
 		}
 		else {
-			try {
+			try (
 				final InputStream fis = new FileInputStream(inFile);
+			) {
 				out.println(new String(getDataFromInputStream(fis)));
 				fis.close();
 				LOGGER.fine("Se recupera el fichero: " + inFile.getName()); //$NON-NLS-1$
 			}
 			catch (final IOException e) {
-				LOGGER.severe(ErrorManager.genError(ErrorManager.ERROR_INVALID_DATA));
+				LOGGER.severe("Error recupendo el fichero '" + inFile.getAbsolutePath() + ": " + e); //$NON-NLS-1$ //$NON-NLS-2$
 				out.println(ErrorManager.genError(ErrorManager.ERROR_INVALID_DATA));
 				return;
 			}
-
-			inFile.delete();
-		}
-	}
-
-	/**
-	 * Elimina del directorio temporal todos los ficheros que hayan sobrepasado el tiempo m&aacute;ximo
-	 * de vida configurado.
-	 * @param config Opciones de configuraci&oacute;n de la operaci&oacute;n.
-	 */
-	private static void removeExpiredFiles() {
-		final File tempDir = ConfigManager.getAfirmaTempDir();
-
-			for (final File file : tempDir.listFiles()) {
-				try {
-					if (file.isFile() && isExpired(file, ConfigManager.getAfirmaTempsTimeout())) {
-						LOGGER.fine("Eliminamos el fichero caducado: " + file.getAbsolutePath()); //$NON-NLS-1$
-						file.delete();
-					}
-				}
-				catch(final Exception e) {
-					// Ignoramos cualquier error suponiendo que el fichero ha sido eliminado por otro hilo
-				}
+			if (!StorageConfig.DEBUG) {
+				inFile.delete();
 			}
-	}
-
-	/**
-	 * Comprueba si un fichero est&eacute; caducado con respecto al tiempo indicado.
-	 * @param file Fichero que se desea comprobar.
-	 * @param expirationTimeLimit Tiempo l&iacute;mite en milisegundos que debe
-	 * transcurrir para considerar caducada una petici&oacute;n.
-	 * @return {@code true} si el fichero ha caducado, {@code false} en caso contrario.
-	 */
-	private static boolean isExpired(final File file, final long expirationTimeLimit) {
-		return System.currentTimeMillis() - file.lastModified() > expirationTimeLimit;
+		}
 	}
 
 	private static final int BUFFER_SIZE = 4096;
